Change the versioning scheme to use a third number
NXT Improved firmware will now use a x.y.z version numbering scheme. The
current version is 1.29.0 and the next one will be 1.29.1.

Previously, the commit hash was included in the custom version. This is
replaced with a build date corresponding to the newest git commit, unless
the SOURCE_DATE_EPOCH environment variable is defined, in which case it
will be used as build date.

When out of git, if no SOURCE_DATE_EPOCH is defined, it will fall back
to the current date.

The weird date parsing code in BtTest is gone.