Fix signed float to int conversion
This can be visible using this test:

    task main()
    {
      float a = -12.34;
      NumOut(0, 32, a);

      char b = a;
      NumOut(0, 24, b);

      Wait(3000);
    }

Thanks to Andreas Weber for reporting the problem and for the test
program.

This was actually fixed in NXT Enhanced Firmware, not sure if it was by
accident or not as the original firmware source code comment suggests that
special casing this conversion could improve speed.

Imported from NXT Enhanced Firmware.