Fix float to integer conversion
When converting a float to an integer, if the integer is not able to
represent the value, the behavior is implementation defined.

The original firmware source code depends on this implementation defined
behavior from the C compiler.

The LEGO MINDSTORMS NXT Executable File Specification does not define
what is the expected behavior for the firmware because it was released
for version 1.03, before float was supported. However it says: "At the
scalar level, data type conversions behave identically to type casts in
ANSI C.".

For NXT Improved Firmware, I decided to follow the behavior of the original
firmware binary, which is:

 - Negative values to unsigned numbers: large positive integer (two's
   complement).
 - Round to nearest for SETOUT instruction, truncate for other
   operations.